requirements.txt
Flask==1.1.2
moviepy==1.0.3
opencv-python==4.5.2.52
Pillow==8.2.0
tensorflow==2.4.1
Werkzeug==1.0.1
